+++ .gitignore
@@ -2,3 +2,4 @@
 NFT_SNiper.egg-info/
 build/
 __pycache__
+extra/
