+++ NFT_Sniper/utils.py
@@ -2,15 +2,19 @@ import json
 import requests
 import math
 import numpy as np
+import pickle 
 
 from PIL import Image
 from io import BytesIO
+from functools import reduce
 
 # Useful Lambas
+remove = lambda s, rm: s.translate({ord(i): None for i in rm}) # string removal
+reduce_df = lambda dfs: reduce(lambda df1, df2: df1.merge(df2, how='outer'), dfs)
 get = lambda url: requests.request("GET", url)
 jtext = lambda url: json.loads(requests.request("GET", url).text)
 maybe_unlist = lambda x: x[0] if len(x) == 1 else x
-# keys_to_global_nms = lambda d: [globals()[k] = v for k, v in d.items()]
+# keys_to_global_nms = lambda d: [globals().update(k, v) for k, v in d.items()]
 
 # Creates an empty dict with the specified type as a default value
 def typedict(keys=None, dtype=int):
@@ -50,3 +54,47 @@ def inorder(x, asc=True):
         while x[i-1] >= x[i]: i += 1
     del x[len(x)-1]
     return True if i == len(x) else False
+
+def is_numpy(x):
+    return x.__class__ in [
+        np.ndarray,
+        np.rec.recarray,
+        np.char.chararray,
+        np.ma.masked_array
+    ]
+
+def is_scalar(x):
+    if is_numpy(x):
+        return x.ndim == 0
+    if isinstance(x, str) or type(x) == bytes:
+        return True
+    if hasattr(x, "__len__"):
+        return len(x) == 1
+    try:
+        x = iter(x)
+    except:
+        return True
+    return np.asarray(x).ndim == 0
+
+def unpickle(fp):
+    with open(fp, 'rb') as f:
+        x = pickle.load(f)
+    return x
+
+def jload(fp):
+    with open(fp, 'rb') as f:
+        x = json.load(f)
+    return x
+
+def loadz(fp, key='arr_0'):
+    x = np.load(fp, allow_pickle=True)
+    if is_scalar(key):
+        return x[key]
+    return {k: v for k,v in x.items() if k in keys}
+
+def dict_up(l):
+    """Wrap up a list of dicts into one dict"""
+    raise NotImplementedError
+
+def flatten_traits(d):
+    return np.concatenate([list(v.keys()) for v in d.values()])
